alods v0.1.0 Alods

Documentation for Alods.

Functions

list_queue()

list_queue() :: [%Alods.Queue.Record{created_at: term, data: term, id: term, last_failure_reason: term, method: term, retries: term, status: term, timestamp: term, updated_at: term, url: term}]

Returns a list of all current notifications that are waiting for delivery

notify_by_get(url, data)

notify_by_get(String.t, map | list) :: {:ok, String.t}

Deliver a notification via an HTTP GET request to the given url the data will be url-encoded and attached as query string to the url

notify_by_post(url, data)

notify_by_post(String.t, map | list) :: {:ok, String.t}

Deliver a notification via an HTTP POST request to the given url. The data will be JSON encoded and be attached to the body

queue_size()

queue_size() :: non_neg_integer

Returns the number of notifications that are waiting for delivery

stress()