alods v0.1.0 Alods.Delivered

This module takes care of starting a DETS store which will hold delivered messages.

Functions

clean_store()

clean_store() :: pid | {pid, reference}

This function will initiate a DETS table clean, meaning it will remove all entries which are older then the configured store time, which by default is 7 days.

clear!()

clear!() :: :ok

Clears the store, WARNING this removes all records!

delete(id)

delete(String.t) :: :ok | {:error, any}

Finds a record by id, and deletes that record

delete!(record)

delete!(%Alods.Record{created_at: term, data: term, delivered_at: term, id: term, method: term, reason: term, retries: term, status: term, timestamp: term, updated_at: term, url: term}) :: :ok

Deletes the given record

find(id)

find(String.t) ::
  {:ok, %Alods.Record{created_at: term, data: term, delivered_at: term, id: term, method: term, reason: term, retries: term, status: term, timestamp: term, updated_at: term, url: term}} |
  {:error, :record_not_found}

Finds a record by id

length()

length() :: number

Get the current amount of record entries in the store

list()

list() :: list

Will return all records in the store

size()

size() :: number

Alias for length/1

start_link()

start_link() :: {:ok, pid}

Starts the Genserver which will own the DETS table and is the only one that has write rights.

store(record)

store(%Alods.Record{created_at: term, data: term, delivered_at: term, id: term, method: term, reason: term, retries: term, status: term, timestamp: term, updated_at: term, url: term}) :: :ok

Stores the given record, updating the delivred at field, resetting the reason, and setting the status to delivered. After successful storing, it will be deleted from the Queue.