alods v0.1.0 Alods.Record

These are the records that are stored in the queue.

Functions

create(params \\ %{})

create(params :: map | list) ::
  {:ok, %Alods.Record{created_at: term, data: term, delivered_at: term, id: term, method: term, reason: term, retries: term, status: term, timestamp: term, updated_at: term, url: term}} |
  {:error, Ecto.Changeset.t}

Use this to create the Alods.Record struct which will validate all options given.

update(record, params)

update(%Alods.Record{created_at: term, data: term, delivered_at: term, id: term, method: term, reason: term, retries: term, status: term, timestamp: term, updated_at: term, url: term}, map) ::
  {:ok, %Alods.Record{created_at: term, data: term, delivered_at: term, id: term, method: term, reason: term, retries: term, status: term, timestamp: term, updated_at: term, url: term}} |
  {:error, any}

Will update a given record

update!(record, params)

update!(%Alods.Record{created_at: term, data: term, delivered_at: term, id: term, method: term, reason: term, retries: term, status: term, timestamp: term, updated_at: term, url: term}, map | list) :: %Alods.Record{created_at: term, data: term, delivered_at: term, id: term, method: term, reason: term, retries: term, status: term, timestamp: term, updated_at: term, url: term}

Will Update a given record and raise if validation failed or something else went wrong