alods v0.1.0 Alods.Store

This module takes care of starting a DETS store.

Functions

clear!()

clear!() :: :ok

Clears the store, WARNING this removes all records!

delete(id)

delete(String.t) :: :ok

find(id)

find(String.t) :: %Alods.Store.Record{data: term, id: term, last_failure_reason: term, method: term, retries: term, status: term, timestamp: term, url: term}

get_pending_entries()

get_pending_entries() :: {:ok, [{String.t, atom, String.t, map, number, atom}]}

Returns all entries of which their timestamp are smaller then the current time

get_work()

length()

length() :: number

Get the current amount of record entries in the store

list()

list() :: list

Will return all records in the store

push(method, url, data)

push(atom, String.t, map) :: {:ok, String.t} | {:error, any}

Pushes a record into the store.

retry_later(record, reason)

retry_later(%Alods.Store.Record{data: term, id: term, last_failure_reason: term, method: term, retries: term, status: term, timestamp: term, url: term}, any) ::
  :ok |
  {:error, any}

size()

size() :: number

Alias for lenth/1

start_link()

start_link() :: {:ok, pid}